converted js to tsx
