+++ app/layout.tsx
@@ -1,0 +1,16 @@
+type RootLayoutProps = {
+  children: React.ReactNode;
+}
+
+export const metadata = {
+  title: 'Next.js',
+  description: 'Generated by Next.js',
+}
+
+export default function RootLayout({ children }: RootLayoutProps) {
+  return (
+    <html lang="en">
+      <body>{children}</body>
+    </html>
+  )
+}
